Create Quicksort 2 - Sorting(HackerRank)
In this challenge, print your array every time your partitioning method finishes, i.e. whenever two subarrays, along with the pivot, are merged together.

The first element in a sub-array should be used as a pivot.
Partition the left side before partitioning the right side.
The pivot should be placed between sub-arrays while merging them.
Array of length  or less will be considered sorted, and there is no need to sort or to print them.