Create Leetcode- 456. 132 Pattern
Here we used a monotonic stack (a monotonic stack is a specialized variation of the traditional stack data structure where the elements are always maintained in a specific order, either strictly increasing or strictly decreasing). I traverse from right to left. Find out the second max value from the stack, as its index is greater (k), and store the max value in the stack, which is j and less than k. Then just needed to find out i which must be less then k value.